+++ lib/voicemail/call_controllers/mailbox_controller.rb
@@ -1,5 +1,8 @@
 module Voicemail
   class MailboxController < ApplicationController
+
+    attr_accessor :new_or_saved
+
     def run
       if mailbox
         play_number_of_messages :new
@@ -11,18 +14,30 @@ module Voicemail
     end
 
     def play_number_of_messages(new_or_saved)
-      get_count(new_or_saved)
+      @new_or_saved = new_or_saved
+      get_count
 
       if @number > 0
-        play config.mailbox.number_before
-        play_numeric @number
-        play config.mailbox["number_after_#{new_or_saved}".to_sym]
+        play_message_count
       else
         play config.messages["no_#{new_or_saved}_messages".to_sym]
       end
     end
 
-    def get_count(new_or_saved)
+    def play_message_count
+      case config.numberic_method
+      when :play_numeric
+        play config.mailbox.number_before
+        play_numeric @number
+        play config.mailbox["number_after_#{new_or_saved}".to_sym]
+      when :ahn_say
+        play config.mailbox.number_before
+        play *sounds_for_number(@number)
+        play config.mailbox["number_after_#{new_or_saved}".to_sym]
+      end
+    end
+
+    def get_count
       @number = storage.send "count_#{new_or_saved}_messages", mailbox[:id]
     end
   end