+++ spec/voicemail/call_controllers/mailbox_play_message_controller_spec.rb
@@ -45,11 +45,8 @@ describe Voicemail::MailboxPlayMessageController do
 
   describe "#intro_message" do
     it "plays the message introduction" do
-      subject.should_receive(:current_message).and_return(message)
-      should_play config.messages.message_received_on
-      subject.should_receive(:play_time).once.with(message[:received], format: config.datetime_format)
-      should_play config.messages.from
-      subject.should_receive(:say_characters).once.with "39335135335"
+      subject.should_receive(:current_message).once.and_return message
+      should_invoke Voicemail::MailboxPlayMessageIntroController, message: message, mailbox: 100, storage: storage_instance
       controller.intro_message
     end
   end