+++ CHANGELOG.md
@@ -7,6 +7,7 @@
   * FEATURE - Allow overriding `MailboxMainMenuController` with your own custom menu class
   * FEATURE - Ability to listen to new OR saved messages, and archive/unarchive at will.
   * FEATURE - Improved passing of a storage mechanism between controllers
+  * FEATURE - Customizable method for using TTS/ahn_say for things like "You have #{x} new messages"
   * CS - Rename config option for when to answer
   * CS - DRY up duplicated `#mailbox_not_found` method
   * CS - Move all the call controllers to their own call_controller directory, to avoid clutter